Move common gpu primitives to backend/gpu
