+++ mlx/backend/gpu/slicing.h
